pyproject.toml
[tool.poetry]
name = "textbot"
version = "0.1.0"
description = "The project is about creating a chatbot that will leverage LLM to respond to queries related to the provided text files"
authors = ["Udit Kumar Chatterjee <quantumudit@gmail.com>"]
license = "NC-SA"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.6"
openai = "^1.12.0"
chromadb = "^0.4.22"
python-dotenv = "^1.0.1"
python-box = "^7.1.1"
pyyaml = "^6.0.1"
streamlit = "^1.31.0"
langchain-community = "^0.0.19"
langchain-openai = "^0.0.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
